+++ braindecode/models/sleep_stager_eldele_2021.py
@@ -44,7 +44,7 @@ class SleepStagerEldele2021(EEGModuleMixin, nn.Module):
         input dimension of the second FC layer in the same.
     n_attn_heads : int
         Number of attention heads. It should be a factor of d_model
-    dropout : float
+    drop_prob : float
         Dropout rate in the PositionWiseFeedforward layer and the TCE layers.
     after_reduced_cnn_size : int
         Number of output channels produced by the convolution in the AFR module.
@@ -82,7 +82,7 @@ class SleepStagerEldele2021(EEGModuleMixin, nn.Module):
         d_model=80,
         d_ff=120,
         n_attn_heads=5,
-        dropout=0.1,
+        drop_prob=0.1,
         activation_mrcnn: nn.Module = nn.GELU,
         activation: nn.Module = nn.ReLU,
         input_window_seconds=None,
@@ -151,10 +151,10 @@ class SleepStagerEldele2021(EEGModuleMixin, nn.Module):
             activation_se=activation,
         )
         attn = _MultiHeadedAttention(n_attn_heads, d_model, after_reduced_cnn_size)
-        ff = _PositionwiseFeedForward(d_model, d_ff, dropout, activation=activation)
+        ff = _PositionwiseFeedForward(d_model, d_ff, drop_prob, activation=activation)
         tce = _TCE(
             _EncoderLayer(
-                d_model, deepcopy(attn), deepcopy(ff), after_reduced_cnn_size, dropout
+                d_model, deepcopy(attn), deepcopy(ff), after_reduced_cnn_size, drop_prob
             ),
             n_tce,
         )